Use a new strategy for sandboxing the recordBug/test_likepy future.
The idea of just removing functions that still contain yield primitives after the
lowerIterators pass doesn't pan out.  The target test case
  studies/madness/aniruddha/madchap/recordBug/test_likepy
still fails, this time with a segfault in the compiler.  Even though the function is
removed, there is still a call to it in the AST.  Duh.

In the new approach, invalid PRIM_YIELD primitives that remain in the tree after
lowerIterators has run are replaced by PRIM_INT_ERROR primitives.  This demonstrates that
the recursive iterator is actually called by the application, so it is actually important
to get the translation correct.

Updated the .bad file to match the (now very stable) runtime error message.

The actual fix is left as a future task.

This commit also improved comments a bit, and clarified the code by providing mnemonic
names for repeated "call->parentExpr" and "call->parentSymbol" expressions.