Adding --no-print-directory to MAKEFLAGS
This seems to be what was breaking the Travis builds.  Previously,
we got this setting from Makefile.base.  Now that we're no longer
including Makefile.base in the top-level directory, sub-make
commands were not getting this setting and we were getting extra
output in the 'make check' command that caused the diff to fail.
Thomas anticipated this, but I wasn't seeing it in my local make
checks which is why I shrugged the suggsetion off.