Move my performance "playground"
My playground originated as doing qthreads performance testing alongside of
fifo and devolved into a testing a bunch of different performance related
things. I believe having the playground around is useful longterm to
investigate performance related things without having to manually kick off
runs. However, it's not useful being on the main performance webpage and is
confusing to most people on the team and almost certainly all people off the
core team.

This just moves it from being at the page to being on it's own that is not
publicly linked.

E.G instead of being chapel.sourceforge.net/perf/chap04, it will be
chapel.sourceforge.net/perf/chap04/playground