Symbolic CHPL_RT_NUM_THREADS_PER_LOCALE values: MAX_PHYSICAL and MAX_LOGICAL.
Add support for two symbolic values for CHPL_RT_NUM_THREADS_PER_LOCALE:
'MAX_PHYSICAL' and 'MAX_LOGICAL'.  These will be interpreted as the
number of physical and logical CPU instances available to the program,
respectively.  Also, add unit tests for these values and as part of
this, introduce a new section of tests/ to contain runtime-specific
tests.