Updating util/README in preparation for the 1.10 release
Broke the previous description of the setchplenv.* scripts into two
sections -- one for quickstart, one for experienced users.

Broke the organization of the README itself into three sections:
novice, experienced user, and developer

Removed mention of the update-sources script, as I think it's time to
retire it (we haven't kept it up-to-date with recent changes and don't
have any sign that it's been used).