Work around lack of kernel scheduling affinity on KNC compute nodes.
The KNC compute nodes we have access to seem not to have kernel affinity
scheduling present.  Either that, or perhaps it's confused: we get masks
with no bits set returned from sched_getaffinity().  To work around this
for now, just don't apply the scheduling affinity limiting in the two
functions that return the number of physical and logical CPUs, even if
the caller passes the flag saying to pay attention to that.  (Note that
we also ignore this flag for all the non-Linux targets at present, so
there is precedent here.)