Refactoring Makefiles to try and clean them up
[trivial, not reviewed]

Mostly, pushed logic relating to third-party into the
third-party Makefile.  Also make the format of the re2
target more symmetric with others to make the "up-to-date"
message print more nicely.  Moved the build of the optional
packages after the compiler and before the runtime, like
other packages.

I'd hoped that by doing this move, I'd only get the
speculative build message when the build was actually
not already done (by putting the install directory in
as a precondition), but I'd forgotten that that would
change the order in which things were built.  Still,
I think this approach is an improvement on the original.