Unload cray-libsci for whitebox testing.
This was previously selectively done when testing the cray compiler (cce) on
whiteboxes. However, cray-libsci's precense started having issues with other
programming environments recently. So, it is time to unload it for all of them.

It is worth noting, internally the Chapel binary release for Cray systems is
built and tested with cray-libsci module loaded.