Remove --no-bounds-checks compopts from "maximal" range tests
This removes the --no-bounds-checks compopts from
  test/trivial/deitz/test_max_range_iteration
  test/types/range/elliot/test_min_and_max_range_iteration

The original intent of having the explicit --bounds-checks and
--no-bounds-checks was to not have to worry about what the test would do
normally vs for --fast testing. The no-bounds-checks versions were expected to
overflow and cause an infinite loop that the test program would detect.
However, since these two tests would result in signed overflow and signed
overflow is undefined in c anything can happen. For most compilers this did
result in infinite loops, but for the cray compiler the tests stopped like you
would expect a regular loop to, which was a little surprising at first but is
fine since the behavior is undefined.

types/range/hilde/maximalRange did not need to change since unsigned
integer overflow is well defined.