Update the man page and README.chplenv for release.
In README.chplenv, CHPL_MODULE_PATH was not mentioned, so I added an "Optionally ..."
section to cover it.  I also mention that "curl" is now a valid setting for
CHPL_AUX_FILESYS, and direct readers to technotes/README.curl for details.

Entries in the ENVIRONMENT section were alphabetized.  The descriptions were not altered
and none added nor deleted.