Housekeeping on CONTRIBUTORS
We've decided to focus the CONTRIBUTORS file here on those who have
contributed to the implementation rather than the language design
itself (such contributors are acknowledged in the language spec
itself).  To that end, removing Mark James and Hans Zima who did not
contribute code that ended up in the release or repository I can
recall or find in the logs.

Unifying to the term "individual contributor" for those who did not
contribute through an official organization or whose affiliation has
been lost.