Fix Coverity issues 1212401 and 1212402
In stmt.cpp, Coverity noticed that there were null checks for two variables that
had already been checked.  While what it complained about was the possibility
that the later dereference was to a null pointer, it was fair to point out that
the variables did not need to be checked again.  I believe that removing the
additional check will resolve these issues.