Fix for Coverity issue 1212292
Coverity correctly pointed out that the variable debug in checkResolved's
isDefinedAllPaths does not change and so the true branch for if (debug) will
never be executed.  This code appears to be solely for debugging purposes and as
such should only be executed intentional by a developer.  To that end, leave the
code present but in a form where it will not be analysized.