Update start_test to print summary of errors after suppressions have been checked.
Previously, the errors were printed to stdout _and_ added to the summary
logfile. Then suppressions were removed from the summary logfile and any
suppressions that did not show as Errors were reported as error, in the summary
logfile _only_.

This updates start_test to print to stdout the same summary that ends up in the
summary logfile. So suppressions that are removed will not show up in either
place, and missing suppressions (tests that have a suppression, but succeed)
will show up as an Error in both places.