Add new type c_string_copy to track owned c_strings.
Formerly, c_string was used to represent low-level C strings without regard to whether
they behave like fully-constructed (owned) objects or like references (unowned).  To
simplify the task of tracking allocations across the runtime/module interface, I added the
new type c_string_copy.

Now, c_string is used to represent unowned strings and c_string_copy is used to represent
owned strings.  To prevent memory leaks, an owned string should be freed before it goes
out of scope (or its ownership transferred to another program entity).  An unowned string
should not be feed.

When a c_string is used to initialize an object (such as a chpl_string)
that is expected to own its data, it must be copied into that object.  When a
c_string_copy is used to initialize such an object, a pointer copy will suffice.
Ownership of the c_string_copy is then transferred to the new object.

Note that a c_string_copy can be coerced to a c_string without breaking this ownership
invariant.  Ownership is retained by the original (c_string_copy) and not by the duplicate
(c_string).  This seems a good place to justify the naming: A c_string_copy is produced by
passing a c_string to the function string_copy().

[This version compiles and runs hello.chpl]

###DETAILS

####compiler/AST/expr.cpp
Moved all of the conversions between chpl_string, c_string and various fundamental types
from the PRIM_CAST clause of CallExpr::codegen() out to module code.
TODO: A PRIM_CAST returning an owned object should probably allocate that object.  This is
not a problem for fundamental types, but is probably a problem for c_string_copy,
chpl_string and records.  (Record casts are handled by compiler-generated cast functions
unless overridden.)

####compiler/AST/primitive.cpp
Added the new type info functions returnInfoStringCopy, and used this to inform the
compiler that the primitives
  string_concat
  string_index
  string_copy
  string_select
return c_string_copy objects.

####compiler/include/symbol.h
####compiler/AST/symbol.cpp
Added the new global initializer symbol gStringCopy.

####compiler/include/type.h
####compiler/AST/type.cpp
Added the new internal type dtStringCopy.
Added some comments suggesting future revisions.

####compiler/passes/resolveIntents.cpp
Added dtStringCopy to the list of types whose blank intent is equivalent to "const in".

####compiler/resolution/functionResolution.cpp
Added a coercion from c_string_copy to c_string.

####modules/internal/ChapelIO.chpl
Changed the return type of chpl_format() to c_string_copy.  These functions always
returned an owned c_string, and now we advertise that in the interface.

####modules/internal/String.chpl
Added an overload for toString() that takes a c_string_copy operand.  The new version
assumes ownership of the passed-in c_string_copy.  There were notations in the code that
we could use a non-allocating form of toString().  Now that task is accomplished.
Added the non-allocating versions of toString() -- taking a c_string_copy and returning a
chpl_string.
Added the c_string_copy to c_string cast used by the added coercion.

Removed the workaround for the string allocation bug in casts between c_strings and
chpl_string fixed before release, and replaced it with code that should not leak.
Some of the casts from fundamental types (e.g. enums and bools) to c_string yielded
an unowned c_string, while others yielded an owned c_string (i.e. a c_string_copy).  The
c_string cases had to be handled special w.r.t. the call to chpl_free_c_string().  Now,
copies are made of the c_string results, while the c_string_copy results are just usurped
buy the resulting chpl_string.

Added back the casts that used to be implemented by PRIM_CAST:
* bool to c_string
* integral to c_string_copy
* real to c_string_copy
* imag to c_string_copy

Revised c_string.substring() to always return a c_string_copy.  The ownership state of the
returned value used to be ambiguous, because if the range was empty it would return a
reference to a literal empty string (a c_string).

####modules/standard/GMP.chpl
chpl_gmp_mpz_get_str() now returns a c_string_copy.

####modules/standard/IO.chpl
qio_encode_to_string() now returns a c_string_copy.
The cast of an ioChar to a c_string now returns a c_string_copy.

####modules/standard/NewString.chpl
remoteStringCopy now returns a c_string_copy.

####runtime/include/chpl-string.h
Added string_from_c_string_copy() and wide_string_from_c_string_copy() to create a string
from a c_string_copy without allocating new memory.  Ownership of the c_string_copy
argument is transferred to the new string.

####runtime/include/chplcast.h
####runtime/src/chplcast.c
Replaced the individual int type to c_string functions with integral_to_c_string_copy().
Replaced the real and imag type to c_string functions with real_to_c_string_copy().

####runtime/include/chplsys.h
chpl_nodeName() actually returns a shared, statically-allocated string, so I changed its
return type to c_string from c_string_copy.

####runtime/src/chpl-string-support.c
Reworked string_select so it always returns a newly-allocated c_string_copy or NULL.

####runtime/src/chpl-string.c
Added string_from_c_string_copy().  This function steals the data owned by the passed-in
c_string_copy, so it creates a string from a c_string_copy without requiring a new
allocation and copy.  This is is used to implement the "non-allocating toString" mentioned
in several TODOs by Sung (which also allowed me to resolve those TODOs).

Added wide_string_from_c_string_copy(), which is the wide version of the above.

Reworked stringMove() so it always reuses the buffer provided.  The use of a zero-length
string to represent an empty c_string_copy is deprecated in favor of a null pointer.

####runtime/src/chplsys.c
Reworked chpl_nodeName() so it always returns an unowned c_string.
The thread safety of returning a shared object implies that it should be initialized only
once (pre node).  Although I don't guarantee that, it's pretty likely.
Using the assumption that the node name does not change in the course of execution of a
program, I moved the initialization inside the namespace==NULL test, so (in the absence of
races) it is, in fact, initialized only once.
If the node name is expected to change during a run, this routine would have to be
reworked using unshared data, and return a c_string_copy.