Some tweaks after first paratest run.
Now the result of a cast of any type to a string-ish thing is c_string_copy.  Adjusted the
memory management in that routine and also in string.write and c_string.write routines.

Made the coercion cast from c_string_copy to c_string return the right type.
Added a writeThis() method for c_string_copy.

Down to 401 failures (std).