A few tweaks to get tests in StringImpl to work.
Added concatenation operators taking a c_string_copy as one argument, so that internal
casts produce a c_string_copy that is then stolen.

The concatenation helper routines now use a c_string_copy, but leak memory somewhere.

Casts from other types to string_rec now go through the c_string_copy version, and the
output of this is stolen.  Note that special handling of Enum and Bool and ioLiteral and
ioEndOfLine and other literal strings is avoided.

Added a workaround to stringMove, to also allocate a new string if the input string is the
empty string.  This might leak memory in general, but the way it is used in NewString
currently does not cause this problem.