Use WARNINGS to enable warning flags in compiler specific makefiles.
Previously whether or not CHPL_DEVELOPER was defined determined if the warning
flags were added for C compiler. Instead, use the WARNINGS variable. This
allows a developers to build with compiler warnings enabled (e.g. when building
the runtime/) without the rest of the CHPL_DEVELOPER settings. Conversely, if
CHPL_DEVELOPER is set, a developer can disable C compiler warning flags by
setting WARNINGS=0 when calling make. The latter is not especially useful (or
recommended).