Changes in the new string implementation, to eliminate memory leaks.
In versions of proc +() taking a c_string_copy operand, it was necessary to delete (the
representation owned by) that operand.  This is consistent with the rule that the
recipient of a c_string_copy argument is expected to take ownership of it.

Changed string_select() and string_index() to return a NULL to represent an empty string.
The current string_rec implementation will not delete a string of zero length (without
extensive rework), so a NULL is returned instead to avoid leaks.  It does not matter
whether the NULL is reference-counted and deleted or never deleted, so regardless how
string_rec handles this return value, the memory management part comes out right.

TODO: Clients of c_str() should perform a NULL test before attempting to evaluate it as a
string literal.  Either the documentation for c_str should state this clearly, or c_str() be
changed to return a literal "" if the underlying representation is NULL.

stringMove() was modified, so it can accept a NULL input string and will return a NULL in
that case.