Update .skipif logic to support "does not contain" operator.
Add new operator, >= (as opposed to <=), which mean <variable> does not contain
the substring <value>. For example, adding this line,
"CHPL_AUX_FILESYS >= lustre", to a skipif means: skip the test if
CHPL_AUX_FILESYS does not contain the substring "lustre".

I am not really happy with the operator, but I don't think it is any more or
less confusing than the <= operator.