Back explicit cast to c_string_copy out of beer programs.
The beer programs were failing because the return type of describeBottles is established
by the cast of bottleNum to "string" in the "then" clause, but (because of the
concatenations) the type of the else clause is c_string_copy and the compiler finds no way
to promote a value of this type to match the "string" type of the "then" clause.

The solution was to add a clause for the c_string_copy --> string coercion in the
canCoerce() function in the compiler.  This allowed me to put back all the "beer" programs
as they were.