Adding mpz_clear() calls to pidigits tests to plug memory leaks
These changes were motivated by tvandoren's BigInt version
which is leak-free, causing him to point out that most of
our existing versions were not (and the hypothesis is that
making them leak-free won't add significant overhead, based
on the difference between his version and mine).

The ultimate version would be one in which BigInt was a record
in order to have it clean up after itself when its destructor
is called.  This is a background TODO that the two of us are
interested in getting done.