Add chmod capability.
This allows the user to call a Chapel-style version of chmod to modify the
current permissions of a file or directory.  I added the interface to the
FileSystem module and the backend support to chpl-file-utils.c/.h, reserving
the name in reservedSymbolNames since it will conflict with the C definition
otherwise.  I also added two tests: one that the function works on both files
and directories, and one that a correct error message is generated when called
on a non-existent file.