Add isLink functionality
This adds the ability to check within Chapel code if a provided name corresponds
to a soft link.  The interface is visible in the FileSystem module, the
implementation itself lies in chpl-file-utils.c.  Also provide two test cases:
one to exercise files and directories that are and are not links, and one to
exercise the case where the link itself does not exist.