Fixing quickstart shells w.r.t. quickstart on Crays problem TMac hit today
Today, TMac started up a Cray Chapel environment building from source
in which CHPL_COMM would default to gasnet/aries or gasnet/gemini due
to being on a Cray, but the CHPL_MEM layer would be set to cstdlib
because of a seemingly innocuous change made to
util/quickstart/setchplenv.* when we made tcmalloc the default.
Because these GASNet conduits require tcmalloc or dlmalloc or the like
to register memory, an (vague) error is generated.

Here I'm addressing this by:

(a) making the quickstart files set CHPL_COMM to 'none' which arguably
    is what we might want people to start with, even on a Cray if
    they're building from source (and is already the default
    everywhere else);

(b) making the quickstart files not set CHPL_MEM.  This is something
    we will probably want to undo if/when we make tcmalloc the default
    at some point in the future.

Either one of these changes should be sufficient by itself.  Both
seemed like a better reflection of what we would want the quickstart
to do when building from source today.

Note that these changes won't affect Cray module users because they
get a different README and tend not to use the setchplenv.* scripts.