Add support for detecting cray-xe and cray-xc platform.
Look for CLEinfo config file, read it, and inspect the NETWORK value. If it is
"gem", set platform to "cray-xe". If it is "ari", set platform to "cray-xc".

If the CLEinfo file does not exist, or if it does not contain a known NETWORK
value, default to the old platform logic.