Do not use parallel execution when calling `make check`.
start_test breaks because the backend C build (inside chpl compiler) calls
make. However, when that make starts up, it assumes it is a submake and
attempts to communicate with a job server that does not exist. There is
probably some way to fix that (one thought is to set .NOTPARALLEL in the
Makefile.exe), but it is simple enough to call regular make with parallel
execution and then make a second call to make check.