Fix execopts and compopts processing for paratest.server
Mike noticed that if you have something like paratest.server -compopts --llvm
-compopts --verify, that only --verify was picked up. The reason is pretty
simple and was that every time a compopts/execopts was found, it overrode the
previous value instead of concatenating. This just updates execopts and
compopts processing to concatenate instead of overriding. For all the other
arguments, overriding makes sense.